+++ src/cppNGS-TEST/Statistics_Test.h
@@ -876,21 +876,5 @@ TEST_CLASS(Statistics_Test)
 		F_EQUAL2(ancestry.eas, 0.0742, 0.001);
 		S_EQUAL(ancestry.population, "AFR");
 	}
-
-
-	void hrdScore()
-	{
-		CnvList cnvs;
-		cnvs.load( TESTDATA("data_in/hrdScore_lst_cnvs.tsv") );
-
-		QCCollection hrd_results = Statistics::hrdScore(cnvs, GenomeBuild::HG19);
-
-		I_EQUAL(hrd_results.value("QC:2000062", true).asInt() , 2);
-		I_EQUAL(hrd_results.value("QC:2000063", true).asInt() , 3);
-		I_EQUAL(hrd_results.value("QC:2000064", true).asInt() , 3);
-
-	}
-
-
 };