+++ doc/GSvar/changelog.md
@@ -1,5 +1,15 @@
 # GSvar change log
 
+### 2022_12-24 (09.01.23)
+
+- variant tab: showing mosaic flag after genotype.
+- small variant search dialog: showing mosaic flag after genotype.
+
+### 2022_12-15 (21.12.22)
+
+- Somatic Report: Format changes and HRD by scarHRD
+- bugfixes
+
 ### 2022_11-75 (14.12.22)
 
 - Filters: adapted filter for 'mosaic' and 'low_mappability' variants.